# openapi.yaml
openapi: "3.1.0"
info:
  title: Food Delivery API
  version: "1.0.0"
  description: REST API for the Food Delivery platform — orders and restaurants.
servers:
  - url: http://localhost:8080
    description: Local (via API Gateway)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    OrderStatus:
      type: string
      enum:
        - PENDING
        - ACCEPTED
        - PREPARING
        - READY_FOR_PICKUP
        - IN_TRANSIT
        - DELIVERED
        - CANCELLED
        - REFUNDED

    OrderItem:
      type: object
      required: [menuItemId, quantity]
      properties:
        menuItemId:
          type: string
          format: uuid
        name:
          type: string
        quantity:
          type: integer
          minimum: 1
        unitPrice:
          type: number
          format: double

    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        restaurantId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        status:
          $ref: "#/components/schemas/OrderStatus"
        totalAmount:
          type: number
          format: double
        stripePaymentIntentId:
          type: string
        deliveryAddress:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateOrderRequest:
      type: object
      required: [restaurantId, items, deliveryAddress]
      properties:
        restaurantId:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItem"
        deliveryAddress:
          type: string

    UpdateOrderStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"

    MenuItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        restaurantId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stockQuantity:
          type: integer
          description: "-1 means unlimited"
        available:
          type: boolean

    Restaurant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        cuisineType:
          type: string
        rating:
          type: number
          format: float
        openNow:
          type: boolean
        menuItems:
          type: array
          items:
            $ref: "#/components/schemas/MenuItem"

    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        details:
          type: string

security:
  - bearerAuth: []

paths:
  # ─── Orders ──────────────────────────────────────────────────────────────────
  /orders:
    post:
      summary: Create a new order
      operationId: createOrder
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized

    get:
      summary: List orders for the authenticated customer
      operationId: listOrders
      tags: [Orders]
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/OrderStatus"
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
        "401":
          description: Unauthorized

  /orders/{orderId}:
    get:
      summary: Get order by ID
      operationId: getOrder
      tags: [Orders]
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found

  /orders/{orderId}/status:
    patch:
      summary: Update order status
      operationId: updateOrderStatus
      tags: [Orders]
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "409":
          description: Invalid status transition

  /orders/{orderId}/cancel:
    post:
      summary: Cancel an order and trigger refund
      operationId: cancelOrder
      tags: [Orders]
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "409":
          description: Order cannot be cancelled in its current state

  # ─── Restaurants ─────────────────────────────────────────────────────────────
  /restaurants:
    get:
      summary: List restaurants
      operationId: listRestaurants
      tags: [Restaurants]
      parameters:
        - in: query
          name: cuisine
          schema:
            type: string
        - in: query
          name: openNow
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: "#/components/schemas/Restaurant"
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer

  /restaurants/{restaurantId}:
    get:
      summary: Get restaurant by ID with full menu
      operationId: getRestaurant
      tags: [Restaurants]
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Restaurant found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
        "404":
          description: Restaurant not found

  /restaurants/{restaurantId}/menu:
    get:
      summary: Get menu items for a restaurant
      operationId: getMenu
      tags: [Restaurants]
      parameters:
        - in: path
          name: restaurantId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"